+++ base.php
@@ -38,8 +38,9 @@ $RequestP   = new RequestParams($request);
 $mcc        = new MCC();
 $menu       = new IPPMenu();
 $utils      = new IPPUtils();
-$language   = $_SESSION['language'] ?? "en-gb";
 
+$langs = $utils->prefered_language(["da","en","da-dk","en-gb"], $_SERVER["HTTP_ACCEPT_LANGUAGE"]);
+$language   = $_SESSION['language'] ?? array_key_first($langs);
 $REQ        = $RequestP->getRequestParams($_SERVER["REQUEST_METHOD"]);
 
 $inline_css = [];