+++ CHANGES.md
@@ -1,2 +1,4 @@
 - Setup of Compliance Administrator as Privileged User
 - Search for Card Details
+- Bugfix: Did not show errorcode on incorrect login
+- Change language of portal to browser-language